[wip] feat. reuse service host client: split PqServiceHostClientLight.ts out of PqServiceHostClient.ts
